added && optimization: seperation of components
