+++ src/app/components/common/index.ts
@@ -28,6 +28,7 @@ export { default as FrequentlyAskedQuestion } from "./FaqComponent/FAQ";
 export { default as Modal } from "./Modal/Modal";
 
 // Conversation Components
-export { default as ChatSidebar } from "./Conversation/chat/Chat-Sidebar";
-export { default as ChatPannel } from "./Conversation/chat/Chat-Pannel";
-export { default as ChatDisplay } from "./Conversation/chat/Chat-Display";
+export { default as ChatSidebar } from "./Conversation/chat/ChatSidebar";
+export { default as ChatPanel } from "./Conversation/chat/ChatPanel";
+export { default as ChatDisplay } from "./Conversation/chat/chat-panel/ChatDisplay";
+export { default as ChatInput } from "./Conversation/chat/chat-panel/ChatInput";
