+++ src/app/store/tenant.store.ts
@@ -146,7 +146,7 @@ export const useTenantStore = create<TenantState>((set) => ({
     try {
       const res = await tenantServices.createTenantRoles(payload);
       set((state) => ({
-        createdRole: res.data, // ✅ res.data is the role object
+        createdRole: res.data,
         roles: [...state.roles, res.data],
         roleSuccess: res.message || "Role successfully created",
         roleLoading: false,
